Implemented the stubbed reload_watches function.
This function allows me to reload the file descriptors from the
filenames specified on the command line at any time in program
execution, instead of just doing it at the beginning. This change was
specifically made so that files that are opened with Vim and other
editors will still be tracked when they are saved.