Changed signals that are being caught.
Previously, the program would trigger on pretty much any event, but this
causes problems for programs like the Python interpreter which will open
a file a bunch of times. Really, we only want to trigger when a file is
modified in some way. This is trickier than it seems at first, because
different programs have different ways of doing this.